+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.8.0] - 2026-09-04
+
 ### Added
 
 - `url` property in Article objects, to link to HTML post document.
@@ -130,7 +132,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added initial gem scaffolding.
 - Added a minimal example site.
 
-[Unreleased]: https://github.com/social-web-foundation/jekyll-activitypub-static/compare/v0.7.6...HEAD
+[Unreleased]: https://github.com/social-web-foundation/jekyll-activitypub-static/compare/v0.8.0...HEAD
+[0.8.0]: https://github.com/social-web-foundation/jekyll-activitypub-static/compare/v0.7.6...v0.8.0
 [0.7.6]: https://github.com/social-web-foundation/jekyll-activitypub-static/compare/v0.7.5...v0.7.6
 [0.7.5]: https://github.com/social-web-foundation/jekyll-activitypub-static/compare/v0.7.4...v0.7.5
 [0.7.4]: https://github.com/social-web-foundation/jekyll-activitypub-static/compare/v0.7.3...v0.7.4